refactor: simplify Anchor, and fix #6473
